==> notebooks/3_Tweets_DataCleaning.ipynb <==
#pip install dask

#import dask.dataframe as dd
import numpy as np
import pandas as pd

#from dask.diagnostics import ProgressBar
#ProgressBar().register()

import nltk
import re
import string
import time
import csv
from tqdm import tqdm

nltk.download('stopwords')
from nltk.corpus import stopwords
cachedStopWords = stopwords.words("english")

## If you are using the data by mounting the google drive, use the following :
# from google.colab import drive
# drive.mount('/content/gdrive',  force_remount=False)

# %cd /content/gdrive/MyDrive/FinalThesis/Bitcoin_Datasets/

#Load Twitter Dataset
start = time.time()
tweet_raw= pd.read_csv("Bitcoin_tweets.csv", encoding = 'utf8')
t = time.time() - start
print('Time taken to load Twitter Data:',t)

text_data = tweet_raw.drop(columns= ['user_name','user_location','user_description','user_created','user_followers','user_friends','user_favourites','user_verified','is_retweet','hashtags','source'])

text_data.dtypes

text_data['date_time'] = pd.to_datetime(text_data['date'], unit='d', origin='1899-12-30')
text_data  = text_data.astype({"text": "string"})

text_data.dtypes

text_data = text_data.drop('date', axis = 1)

text_data.head(10)

# Data Cleaning

a)	Convert tweets to lowercase. D

b)	Expand words with contractions.

c)	Removed duplicate tweets based on unique id. NA

d)	Remove stopwords from the tweet. D

e)	Remove punctuation and special characters. D

f)	Remove user mentions that are a word starting with the @ sign. D

g)	Replace over 2 dots and multiple spaces with a single space. D

h)	Remove dollar symbols from the text like $BTC and $Bitcoin. D

i)	Remove the hashtag sign from the text. D

j)	Remove non-English tweets. D

k)	Remove URLs from tweets. D
  

def clean_text_data(tweet):
  # Transform to lower case
  tweet = tweet.lower()
  #Remove url, square and round brackets, mentions
  tweet = re.sub(r"(http[s]?\://\S+)|([\[\(].*[\)\]])|([@]\S+)", "",tweet)
  #Remove numbers
  tweet = re.sub(r'\d', '', tweet) #reconsider this step
  #Remove stop words
  tweet = ' '.join([word for word in tweet.split() if word not in cachedStopWords  and  word.isascii()])
  #Remove Punctuations
  tweet = tweet.translate(str.maketrans('', '','!"#$%&.\'()*+,-/:;<=>?@[\]^_`{|}~'))
  #Remove extra spaces
  tweet = re.sub('[\s]+', ' ', tweet)

  return tweet

#print (clean_text_data('Blue Ridge Bank shares ------- halted by NYSE  ~ after #bitcoin .......ATM announcement https://t.co/xaaZmaJKiV @MyBlueRidgeBank… https://t.co/sgBxMkP1SI'))

#text_data = text_data.head(20000)

text_data.shape

text_data = text_data.dropna()

text_data.shape

start = time.time()
tqdm.pandas(desc="Cleaning Data........")
text_data["clean_text"] = text_data.text.progress_apply(lambda x: clean_text_data(x))
t = time.time() - start
print("Time taken to clean textual data:",t)

text_data.head(10)

text_data = text_data[['date_time','clean_text']]

clean_data = text_data.dropna()

clean_data.shape

clean_data['date_time'] = clean_data['date_time'].dt.strftime('%Y-%m-%d %H:%M:%S')

clean_data.head(10)

today_date = time.strftime("%d-%m-%Y")
excel_filename = "clean_data_"+today_date +".csv"
print(excel_filename)
clean_data.to_csv(excel_filename, index=False, header=True)
