==> notebooks/4_SentimentScore_FinBERT.ipynb <==
!pip install transformers==4.36.2

import multiprocessing
multiprocessing.cpu_count()

import numpy as np
import pandas as pd
import time

# Fetch Clean data
## If you are using the data by mounting the google drive, use the following :
from google.colab import drive
drive.mount('/content/gdrive',  force_remount=False)

%cd /content/gdrive/MyDrive/FinalThesis/Final_Data/

#import torch
#from torch.utils.data.dataset import Dataset
import scipy
#from transformers import pipeline
#from transformers import AutoTokenizer, AutoModelForSequenceClassification
from transformers import BertTokenizer, BertForSequenceClassification, AdamW, BertConfig

#Load Twitter Dataset
start = time.time()
tweet_clean = pd.read_csv("clean_data_20-01-2024.csv")
t = time.time() - start
print('Time taken to load Twitter Clean Data:',t)

tweet_clean.dtypes

tweet_clean['date_time'] = pd.to_datetime(tweet_clean['date_time'])
tweet_clean['clean_text'] = tweet_clean['clean_text'].astype(str)

tweet_clean.head()

tweet_clean.shape

dataByMonth= tweet_clean[(tweet_clean['date_time'].dt.month == 8)& ((tweet_clean['date_time'].dt.day >= 26) & (tweet_clean['date_time'].dt.day <= 31))]
dataByMonth = dataByMonth.dropna()


#ddf.npartitions
#ddf=tweet_clean.head(100)
ddf=dataByMonth

finbert_tokenizer = BertTokenizer.from_pretrained('ProsusAI/finbert')
finbert_model = BertForSequenceClassification.from_pretrained('ProsusAI/finbert')

labels = finbert_model.config.id2label.values()
from scipy import special
def getSenimentScore(clean_text):
  clean_data = finbert_tokenizer(clean_text, padding=True, truncation=True, return_tensors='pt')
  logits = finbert_model(**clean_data).logits
  scores = {k: v for k, v in zip(labels, scipy.special.softmax(logits.detach().numpy().squeeze()))}
  return scores

#compound_score = (-scores['negative'] + scores['positive'])

from tqdm import tqdm

start = time.time()
print(start)
tqdm.pandas(desc='Calculating Sentiment Score ---')
ddf['compound_score'] = ddf['clean_text'].progress_apply(lambda x: getSenimentScore(x))
t = time.time() - start
print("Time taken to calculate FinBert sentiment score :",t)

ddf.head()

today_date = time.strftime("%d-%m-%Y")
excel_filename = "finbert_compound_score_"+today_date+".csv"
print(excel_filename)
ddf.to_csv(excel_filename, index=False, header=True)
