==> notebooks/6_Aggregate_Calculate_SentimentScore.ipynb <==
import pandas as pd
import glob
import os

## If you are using the data by mounting the google drive, use the following :
from google.colab import drive
drive.mount('/content/gdrive',  force_remount=False)
%cd /content/gdrive/MyDrive/FinalThesis/Final_Data

# data = pd.read_csv('finbert_compound_score_fin.csv')
data = pd.read_csv('roberta_compound_score_fin.csv')

data.shape

#data.head()

data = data.sort_values(by='date_time')
data_f = data.drop('clean_text', axis = 1)
data_f['date'] = pd.to_datetime(data_f['date_time'], format='%Y-%m-%d').dt.floor('D')
data_f = data_f.drop(['date_time'], axis=1)
data_g = data_f.groupby('date').mean()
data_c = data_g.reset_index()

#data.head()

#data_f.dtypes

data_c['compound_score'] = (-data_c['negative'] + data_c['positive']) / (data_c['neutral'] + data_c['negative'] + data_c['positive'])

data_c =data_c.round(4)

#data_c.to_csv("finbert_aggregated.csv", index=False)
data_c.to_csv("roberta_aggregated.csv", index=False)
