==> notebooks/7_Merging_Datasets.ipynb <==
import pandas as pd

## If you are using the data by mounting the google drive, use the following :
from google.colab import drive
drive.mount('/content/gdrive',  force_remount=False)

%cd /content/gdrive/MyDrive/FinalThesis/Final_Data/

data_bitcoin = pd.read_csv('bitcoin_prices.csv')

data_bitcoin.head()

data_bitcoin.columns = ['date', 'open', 'high', 'low', 'close', 'adj_close', 'volume']

data_bitcoin['date'] = pd.to_datetime(data_bitcoin['date'], format='%Y-%m-%d').dt.floor('D')

year = 2021
from_month = 2
to_month = 8
data_bitcoin_fil= data_bitcoin.loc[(data_bitcoin['date'].dt.year == year) & ((data_bitcoin['date'].dt.month >= from_month) & (data_bitcoin['date'].dt.month <= to_month))]

###############################################

tweet_roberta_score = pd.read_csv("roberta_aggregated.csv")

tweet_roberta_score['date'] = pd.to_datetime(tweet_roberta_score['date'], format='%Y-%m-%d').dt.floor('D')

tweet_roberta_score.dtypes

data_r = pd.merge(data_bitcoin_fil, tweet_roberta_score, on="date", how="outer")

data_r = data_r.fillna(0)

data_r.head()

data_final_r = data_r[['date', 'close', 'compound_score']]

data_final_r.head()

%cd /content/gdrive/MyDrive/FinalThesis/Final_Data/

data_final_r.to_csv("roberta_bitcoin_merged.csv", index=False)

##############################################################

tweet_finbert_score = pd.read_csv("finbert_aggregated.csv")

tweet_finbert_score['date'] = pd.to_datetime(tweet_finbert_score['date'], format='%Y-%m-%d').dt.floor('D')

tweet_finbert_score.dtypes

data_f = pd.merge(data_bitcoin_fil, tweet_finbert_score, on="date", how="outer")

data_f = data_f.fillna(0)

data_f.head()

data_final_f = data_f[['date', 'close', 'compound_score']]

data_final_f.head()

%cd /content/gdrive/MyDrive/FinalThesis/Final_Data/

data_final_f.to_csv("finbert_bitcoin_merged.csv", index=False)
