==> notebooks/11_Comparision_of_BERTs.ipynb <==

import numpy as np
import pandas as pd

import nltk
import re
import string
import time
import csv
import scipy
nltk.download('stopwords')
from nltk.corpus import stopwords
cachedStopWords = stopwords.words("english")
from transformers import BertTokenizer, BertForSequenceClassification, AdamW, BertConfig
from transformers import RobertaTokenizer, RobertaForSequenceClassification


def clean_text_data(tweet):
  # Transform to lower case
  tweet = tweet.lower()
  #Remove url, square and round brackets, mentions
  tweet = re.sub(r"(http[s]?\://\S+)|([\[\(].*[\)\]])|([@]\S+)", "",tweet)
  #Remove numbers
  tweet = re.sub(r'\d', '', tweet) #reconsider this step
  #Remove stop words
  tweet = ' '.join([word for word in tweet.split() if word not in cachedStopWords  and  word.isascii()])
  #Remove Punctuations
  tweet = tweet.translate(str.maketrans('', '','!"#$%&.\'()*+,-/:;<=>?@[\]^_`{|}~'))
  #Remove extra spaces
  tweet = re.sub('[\s]+', ' ', tweet)

  return tweet

finbert_tokenizer = BertTokenizer.from_pretrained("ProsusAI/finbert")
finbert_model = BertForSequenceClassification.from_pretrained("ProsusAI/finbert")

labels = finbert_model.config.id2label.values()
def getSenimentScoreByFinBERT(clean_text):
  clean_data = finbert_tokenizer(clean_text, padding=True, truncation=True, return_tensors='pt')
  logits = finbert_model(**clean_data).logits
  scores = {k: v for k, v in zip(labels, scipy.special.softmax(logits.detach().numpy().squeeze()))}

  return scores

 positive_sent = clean_text_data('Bullish on Bitcoin! sats stacking and feeling confident about the future of this innovative tech. #Bitcoin #Cryptocurrency')

negative_sent = clean_text_data('#Bitcoin down another 10% today. ‍♂️  This volatility is insane! Who can trust a currency that swings wildly like this?  More like a digital rollercoaster than a store of value.  #cryptocrash  #bubblebursting')
negative_sent

result_1 = getSenimentScoreByFinBERT(negative_sent)
result_1

result_2 = getSenimentScoreByFinBERT(positive_sent)
result_2

roberta_tokenizer = RobertaTokenizer.from_pretrained("cardiffnlp/twitter-roberta-base-sentiment-latest")
roberta_model = RobertaForSequenceClassification.from_pretrained("cardiffnlp/twitter-roberta-base-sentiment-latest")

labels = roberta_model.config.id2label.values()
from scipy import special
def getSenimentScorebyTwRoberta(clean_text):
  clean_data = roberta_tokenizer(clean_text, padding=True, truncation=True, return_tensors='pt')
  logits = roberta_model(**clean_data).logits
  scores = {k: v for k, v in zip(labels, scipy.special.softmax(logits.detach().numpy().squeeze()))}
  return scores

result_3 = getSenimentScorebyTwRoberta(negative_sent)
result_3

result_4 = getSenimentScorebyTwRoberta(positive_sent)
result_4

df = pd.DataFrame(data = {'FinBert_Neg': result_1, 'FinBert_Pos': result_2, 'TwRoberta_Neg': result_3, 'TwRoberta_Pos': result_4} )
df
